pyproject.toml
[project]
name = "obsidian-graph"
version = "0.1.2"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [ "toml==0.10.2", "networkx", "pyyaml",]

[tool.setuptools.package-data]
obsidian_graph = [ "config.yaml",]

[tool.pytest.ini_options]
testpaths = [ "tests",]
pythonpath = [ "src",]
